refactor: split files into {modal,normal,insert}.go
